requirements.txt
flask>=3.0
gunicorn>=22.0
openpyxl>=3.1
python-dotenv>=1.0
requests>=2.31
google-api-python-client>=2.100
google-auth>=2.23
google-auth-oauthlib>=1.1
